[WIP] Move interval generation to subpackage to prepare support for multiple loaders
