+++ intervals/digital_ocean_function.py
@@ -1,0 +1,3 @@
+def generate_mute_time() -> int:
+    # TODO: Implement this function.
+    return 1
