+++ intervals/intervals.py
@@ -1,0 +1,7 @@
+from config import settings
+
+from . import local
+
+
+def generate_mute_time() -> int:
+    return local.generate_mute_time()
