+++ intervals/time_display_converter.py
@@ -1,8 +1,3 @@
-import random
-
-from ayumi import Ayumi
-from config import settings
-
 
 _WEEKS_IN_MINUTES = 10080
 _DAYS_IN_MINUTES = 1440
@@ -45,28 +40,3 @@ def convert_minutes_to_display_str(minutes: int, granularity=2) -> str:
     start, _, end = ', '.join(result[:granularity]).rpartition(',')
     return start + " and" + end if start else end
 
-
-def generate_mute_time() -> int:
-    # Load the list of intervals used to determine mutes.
-    intervals = settings.get("intervals")
-    Ayumi.debug("Loaded {count} intervals.".format(count=int(len(intervals))))
-
-    # First, select an interval to load
-    # Create the bounds tuples and their respective weights
-    bounds = [interval["bound"] for interval in intervals]
-    weights = [int(interval["weight"]) for interval in intervals]
-    interval = random.choices(bounds, weights=weights, k=1)[0]
-
-    # From the interval, randomly select a time.
-    # The interval is a Tuple[lower_bound: str, upper_bound: str]
-    lower_bound = convert_interval_str_to_minutes(interval["lower"])
-    upper_bound = convert_interval_str_to_minutes(interval["upper"])
-
-    mute_duration = random.randint(lower_bound, upper_bound)
-    Ayumi.debug(
-        "Selected mute duration: ({mute_duration}) from lower bound: ({lower_bound}) and upper bound: ({upper_bound}).".format(
-            mute_duration=convert_minutes_to_display_str(mute_duration),
-            lower_bound=convert_minutes_to_display_str(lower_bound),
-            upper_bound=convert_minutes_to_display_str(upper_bound)))
-
-    return mute_duration
