+++ main.py
@@ -1,13 +1,15 @@
 import random
 import sys
 
-import intervals
+import requests
 
 from ayumi import Ayumi
 from config import settings
 from datetime import datetime, timedelta, timezone
 from discord import Intents, Message
 from discord.ext.commands import Bot
+from intervals import intervals
+from intervals import time_display_converter as tdc
 
 # TODO: Add support for Cogs/Extensions
 
@@ -155,63 +157,67 @@ async def on_message(message: Message):
 
     target_users = target_users[:5]  # Trim to only support up to 5 tags at most currently.
 
-    # Start processing an action for each target_user (either the message author, or all tagged in the message).
-    for target_user in target_users:
-        mute_duration = intervals.generate_mute_time()  # Mute duration is in minutes (as int)
-        mute_duration_display_str = intervals.convert_minutes_to_display_str(mute_duration)
-        Ayumi.info(
-            "Generated mute time: {mute_time} ({mute_time_in_minutes} minutes).".format(
-                mute_time=mute_duration_display_str,
-                mute_time_in_minutes=mute_duration))
+    # Because requests may take long, send a typing indicator so the author knows processing is occuring.
+    async with message.channel.typing():
 
-        target_is_self = (target_user == message.author)
-        Ayumi.debug(f"Target for mute is self: {target_is_self}")
+        # Start processing an action for each target_user (either the message author, or all tagged in the message).
+        for target_user in target_users:
+            mute_duration = intervals.generate_mute_time()  # Mute duration is in minutes (as int)
 
-        # TODO: Log which role on the author is safe.
-        if not set([role.id for role in target_user.roles]).isdisjoint(safe_roles):
-            Ayumi.info("Message {message_id} for user {author_name} has a safe role, ignoring.".format(
-                message_id=message.id,
-                author_name=target_user.name))
+            mute_duration_display_str = tdc.convert_minutes_to_display_str(mute_duration)
+            Ayumi.info(
+                "Generated mute time: {mute_time} ({mute_time_in_minutes} minutes).".format(
+                    mute_time=mute_duration_display_str,
+                    mute_time_in_minutes=mute_duration))
 
-            safe_messages = guild_settings.get("safe_messages_self", [DEFAULT_SAFE_MESSAGE_SELF]) \
-                if target_is_self else guild_settings.get("safe_messages_other", [DEFAULT_SAFE_MESSAGE_OTHER])
-            safe_message_to_use = random.choice(safe_messages)
-            await message.reply(safe_message_to_use.format(
-                safe_user_name=target_user.display_name,
-                mute_duration_display_str=mute_duration_display_str))
-            continue
+            target_is_self = (target_user == message.author)
+            Ayumi.debug(f"Target for mute is self: {target_is_self}")
 
-        # Calculate the time that the user will be unblocked.
-        current_time = datetime.now(timezone.utc)  # UTC time
-        Ayumi.debug("The current time is: {current_time}".format(current_time=current_time.strftime("[UTC] %c")))
+            # TODO: Log which role on the author is safe.
+            if not set([role.id for role in target_user.roles]).isdisjoint(safe_roles):
+                Ayumi.info("Message {message_id} for user {author_name} has a safe role, ignoring.".format(
+                    message_id=message.id,
+                    author_name=target_user.name))
 
-        mute_duration_as_delta = timedelta(minutes=mute_duration)
+                safe_messages = guild_settings.get("safe_messages_self", [DEFAULT_SAFE_MESSAGE_SELF]) \
+                    if target_is_self else guild_settings.get("safe_messages_other", [DEFAULT_SAFE_MESSAGE_OTHER])
+                safe_message_to_use = random.choice(safe_messages)
+                await message.reply(safe_message_to_use.format(
+                    safe_user_name=target_user.display_name,
+                    mute_duration_display_str=mute_duration_display_str))
+                continue
 
-        # TODO: Temporary block due to the Discord timeout being limited to 28 days.
-        if mute_duration_as_delta > timedelta(days=28):
-            mute_duration_as_delta = timedelta(days=28)
-            Ayumi.warning(
-                "WARNING: Generated a mute time above 28 days, trimming to 28 days. Check your configuration!")
+            # Calculate the time that the user will be unblocked.
+            current_time = datetime.now(timezone.utc)  # UTC time
+            Ayumi.debug("The current time is: {current_time}".format(current_time=current_time.strftime("[UTC] %c")))
 
-        unmute_time = current_time + mute_duration_as_delta
-        Ayumi.debug("The user will be unmuted at: {unmute_time}".format(unmute_time=unmute_time.strftime("[UTC] %c")))
+            mute_duration_as_delta = timedelta(minutes=mute_duration)
 
-        await target_user.timeout(mute_duration_as_delta, reason="Timed out via Gacha for {duration}.".format(
-            duration=mute_duration_display_str))
-        Ayumi.info(
-            "Timed out user ({user_name}, {user_display_name}) until {unmute_time}".format(
-                user_name=target_user.name,
-                user_display_name=target_user.display_name,
-                unmute_time=unmute_time.strftime("[UTC] %c")
-            ))
+            # TODO: Temporary block due to the Discord timeout being limited to 28 days.
+            if mute_duration_as_delta > timedelta(days=28):
+                mute_duration_as_delta = timedelta(days=28)
+                Ayumi.warning(
+                    "WARNING: Generated a mute time above 28 days, trimming to 28 days. Check your configuration!")
 
-        mute_messages = guild_settings.get("mute_messages_self", [DEFAULT_MUTE_MESSAGE_SELF]) \
-            if target_is_self else guild_settings.get("mute_messages_other", [DEFAULT_MUTE_MESSAGE_OTHER])
-        mute_message_to_use = random.choice(mute_messages)
-        await message.reply(
-            mute_message_to_use.format(
-                muted_user_name=target_user.display_name,
-                mute_duration_display_str=mute_duration_display_str))
+            unmute_time = current_time + mute_duration_as_delta
+            Ayumi.debug("The user will be unmuted at: {unmute_time}".format(unmute_time=unmute_time.strftime("[UTC] %c")))
+
+            await target_user.timeout(mute_duration_as_delta, reason="Timed out via Gacha for {duration}.".format(
+                duration=mute_duration_display_str))
+            Ayumi.info(
+                "Timed out user ({user_name}, {user_display_name}) until {unmute_time}".format(
+                    user_name=target_user.name,
+                    user_display_name=target_user.display_name,
+                    unmute_time=unmute_time.strftime("[UTC] %c")
+                ))
+
+            mute_messages = guild_settings.get("mute_messages_self", [DEFAULT_MUTE_MESSAGE_SELF]) \
+                if target_is_self else guild_settings.get("mute_messages_other", [DEFAULT_MUTE_MESSAGE_OTHER])
+            mute_message_to_use = random.choice(mute_messages)
+            await message.reply(
+                mute_message_to_use.format(
+                    muted_user_name=target_user.display_name,
+                    mute_duration_display_str=mute_duration_display_str))
 
 
 def main():
